Raise Exception everywhere we self.close in _do_ssl_connect, otherwise async code may never complete execution
Summary: In our async code, we do the write then switch to the parent, expecting to be called back to the child greenlet. But if _do_ssl_connect fails, the socket gets closed and we return from the tornado code without ever scheduling another callback, which prevents that greenlet from completing execution.

Test Plan: tested locally

Reviewers: adam, david, jack, kevin

Differential Revision: https://phab.wish.com/D37619